requirements.txt
beautifulsoup4==4.11.1
pandas==1.4.2
protobuf==3.20.1
requests==2.27.1
google-cloud-storage>=1.41.0
pyarrow>=5.0.0
lxml==4.8.0
openpyxl==3.0.10
numpy==1.21.6